+++ machine/mtrap.h
@@ -33,7 +33,7 @@ extern uintptr_t first_free_paddr;
 extern uintptr_t mem_size;
 extern uintptr_t num_harts;
 extern volatile uint64_t* mtime;
-extern volatile uint16_t* plic_priorities;
+extern volatile uint32_t* plic_priorities;
 extern size_t plic_ndevs;
 
 typedef struct {
@@ -50,9 +50,9 @@ typedef struct {
   uint64_t scycle_delta;
   uint64_t sinstret_delta;
 
-  volatile uint16_t* plic_m_thresh;
+  volatile uint32_t* plic_m_thresh;
   volatile uintptr_t* plic_m_ie;
-  volatile uint16_t* plic_s_thresh;
+  volatile uint32_t* plic_s_thresh;
   volatile uintptr_t* plic_s_ie;
 } hls_t;
